Airport container
