+++ src/containers/Airport/index.ts
@@ -1,0 +1,3 @@
+import Airport from './Airport';
+
+export default Airport;
